requirements.txt
asgiref==3.7.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
decorator==4.4.2
Django==5.0.2
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
gTTS==2.5.1
idna==3.6
imageio==2.34.1
imageio-ffmpeg==0.5.1
moviepy==1.0.3
numpy==1.26.4
pillow==10.3.0
proglog==0.1.10
PyJWT==2.8.0
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
sqlparse==0.4.4
tqdm==4.66.4
urllib3==2.2.0
